backend/requirements.txt
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
geoalchemy2>=0.14.3
alembic>=1.13.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
eval-type-backport>=0.2.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
anthropic>=0.18.0
python-multipart>=0.0.9
weasyprint>=61.0
jinja2>=3.1.3
bcrypt>=4.1.0
geopandas>=0.14.0
shapely>=2.0.0
pyogrio>=0.7.2

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.19.0
